Update for Algo about handshakes using nested loops
Created by Gabe on 06/04/2023.
Update details:
This is the second iteration because the loops solution was not using a long long int and apparently the given one or more sets of input data, consistent with problem-specific constraints were validating it as not a complete solution

Logic part:
The formula calculates the number of distinct pairs that can be formed from a set of n elements.
For example, if n is 4, the formula would give us: 4*(4-1)/2 = 6, which means there are 6 distinct pairs that can be formed from the set of 4 elements, such as (1,2), (1,3), (1,4), (2,3), (2,4), and (3,4).

Note:
Both iterations are considered complete solutions and give max points now when evaluated.