Algo designed to calculate the number of cuts required to cut a pice of paper in 1x1 squares
Created by Gabe on 07/04/2023.

Algo designed to calculate the number of cuts required to cut a pice of paper in 1x1 squares

Logic:
The program reads N and M representing the number of cuts to make than calculates the cuts using the formula: (n - 1) + 1LL * n * (m - 1)//LL is there as a suffix to ensure the program is avoiding  to avoid overflow errors, than is displaying S to the console.