Update #2617 Taieturi.cpp
Update on 08.04.2023 of the exercise to get max score.

Note: in both snippets, the LL suffix is used to ensure that the calculation is done using long long arithmetic, which can prevent potential integer overflow issues.

Also both ver1 and ver2 have the same functionality but ver1 is more explicit, and ver2 is more simplified/faster due to not using a variable to store the formula, yet the performance speed increase is neglectable.