Algo about finding out the time necessary for a given number of cats to eat a given number of mice
Created by Gabe on 13/04/23.

Logic:
This program calculates and outputs the time required for a given number of cats to eat a given number of mice, assuming each cat can eat a fixed number of mice in one minute. It takes two inputs, the number of mice that one cat can eat in one minute and the total number of mice to be eaten, and outputs the time in minutes needed for all the mice to be eaten.