Algo about finding out in how many years will the ratio of Gigel's age to his father's will be equal to c/d, where c an d dare given
Created by Gabe on 14/04/23.

Logic:
The program calculates how many years it will take for the ratio of ages between Gigel and his father to become equal to a given value (c/d). It uses the formula (a * d - b * c) / (c - d) to find the number of years (x), and then displays the result with 2 decimal places using cout.