Algo about finding the sum of second last digit and last digit of a number than output the result
Created by Gabe on 15/04/23.

Logic:
The program reads an integer from the keyboard, extracts the second last digit and the last digit of the number, adds them together and then outputs the result to the console.

In our case with the example 2012
it will take 1(second last digit) and 2(last digit), add them together and output the result 3.