Algo about reading a number from keyboard and than lifting it to the power of 16
Created by Gabe on 16/04/23.

Logic:
I did it in 2 versions, the first one takes variable a and reassigns its value to itself 4 times resulting to a^16
In the second one aka ver2:
It takes the variable a, stores it in another variable named result than takes the said variable and multiplies it to variable a's value in a loop until i=16 each time result will get a new value with is multiplaying itself with a and resulting in the end with the same result as ver1 with is a^16 since this is the end condition for the loop , than displays result to console and terminates the program.