Algo about calculating the square root of a number and display the result with its decimals
Created by Gabe on 18/04/23.

Logic:
This program is very similar to the previous one I made just before this one, the same logic applies with the single difference being the data type of the stored variables.
The n variable is read from keyboard than sqrt function from the cmath library is used on the n variable in the intermediary variable r, afterwards the resulted value of the r variable is displayed to the console together with its decimals as the data type is double now.