Algo about getting the first digit out of a decimal
Created by Gabe on 20/04/23.

Logic:
The program takes variable a with a double data type to include the decimals of the number than static casts the variable to int, with can be done both by static_cast<int> or (int) before the desired variable conversion, than it calculates the value into getting the first digit for the decimal and outputs only the first digit of the decimal as a result to the console.

static_cast<int> is considered more precise way of converting data types of a vriable and it returns an error if it fails so is considered the better way to do it, I still chose to use (int) because I wanted to try it out and bc is easier to read :)