Algo about reading a natural number and using a math formula on it than displaying the result as integer
Created by Gabe on 21/04/23.

Logic:
A natural number is read from the keyboard to assign value to variable a, than the integer part of the desired result is calculated by passing the variable a through the variables  (x,n,m,p,y as they contain the math formula) than the final result after passing through the entire mathematical expression, and converted to integer is displayed to the console as variable y.