Algo about finding the value of a second angle
Created by Gabe on 22/04/23.

Logic:
The program takes an integer input x as the degrees of one of the angles between a segment and a straight line. Than the formula 180-x is used to calculate the y variable value, as requested by the requirement to be found, than the resulted value of y is displayed to the console.

Note:
To be able to solve this algorithm you must know about Euclidean geometry as this is his Triangle Sum Theorem.

The formula y = 180 - x, is expressing the relationship between the angles in a triangle, from applying the Triangle Sum Theorem.