Algo about finding the area of the regular polygon
Created by Gabe on 23/04/23.

Logic:
The program uses the formula " A = (n * l^2) / (4 * tan(pi/n)) "  to calculate the area of a regular polygon based on its number of sides and the length of its sides. It then outputs the result to the console with two decimal points.

Note explaining the formula:
A = (n * l^2) / (4 * tan(pi/n))
The formula involves the tangent function and is based on dividing the polygon into isosceles triangles.