Algo about calculating the set number of an integer using Gauss formula
Created by Gabe on 24/04/23.

Logic:
The program uses Gauss formula " m = ceil((-1 + sqrt(1 + 8*n))/2) " to calculate the set number m that n belongs to. It then outputs the result to the console.

Note explaining the formula:
m = ceil((-1 + sqrt(1 + 8*n))/2)

The formula, uses sqrt to find the square root of 1 + 8*n, and then subtracts 1 from that value. Then divides this value by 2 and takes the ceiling of the result using ceil to get m.