Algo about finding the even numbers in a series of numbers read from keyboard
Created by Gabe on 14/04/24.

Logic:
The program initializes a variable k to count even numbers and a variable x to store all the input number/s

The program enters in a validated as true while loop that continues indefinity until the user enters 0 to exit the loop.
Outside the loop, the program prompts the user to enter a number.
Within the loop the program reads the input number or numbers into variable x
Within the loop the program also checks if any of the entered number/s is 0, if so it breaks out of the loop ending the input process.
Within the loop the program will check for even numbers, if the entered number is not 0, the program checks if it is even by using the modulo operator and the logic that if the number is divided by 2 and the remainder of it is 0, it therefore means that x is even "x % 2 == 0"

If a number is found as being even, then the program increments the counter of k by 1 for each even number.

After the user inputs 0 to terminate, the program checks if any even numbers were entered, if K is still 0, it means no even numbers were found, so it outputs outputs "NU EXISTA" (which means "DOES NOT EXIST" in Romanian).it outputs the count of even numbers.

The program effectively counts the number of even numbers entered by the user until they decide to stop by entering 0, and then it displays the count.

Note:
This is a good example of a base for a program that could be useful to filter and count specific elements in a large dataset.